micro-osd.sh: make shellcheck happier

+++ micro-osd.sh
@@ -1,3 +1,4 @@
+#!/bin/bash
 #
 #    Copyright (C) 2013,2014 Loic Dachary <loic@dachary.org>
 #
@@ -22,14 +23,14 @@ DIR=${1}
 
 # reset
 pkill ceph || true
-rm -rf ${DIR}/*
-LOG_DIR=${DIR}/log
-MON_DATA=${DIR}/mon
-MDS_DATA=${DIR}/mds
-MOUNTPT=${MDS_DATA}/mnt
-OSD_DATA=${DIR}/osd
-RGW_DATA=${DIR}/radosgw
-mkdir ${LOG_DIR} ${MON_DATA} ${OSD_DATA} ${MDS_DATA} ${MOUNTPT} ${RGW_DATA}
+rm -rf "${DIR:?}"/*
+LOG_DIR="${DIR}/log"
+MON_DATA="${DIR}/mon"
+MDS_DATA="${DIR}/mds"
+MOUNTPT="${MDS_DATA}/mnt"
+OSD_DATA="${DIR}/osd"
+RGW_DATA="${DIR}/radosgw"
+mkdir "${LOG_DIR}" "${MON_DATA}" "${OSD_DATA}" "${MDS_DATA}" "${MOUNTPT}" "${RGW_DATA}"
 MDS_NAME="Z"
 MON_NAME="a"
 MGR_NAME="x"
@@ -93,26 +94,26 @@ EOF
 
 launch_mon() {
     ceph-mon --id ${MON_NAME} --mkfs --keyring /dev/null
-    touch ${MON_DATA}/keyring
+    touch "${MON_DATA}/keyring"
     ceph-mon --id ${MON_NAME}
 }
 
 launch_osd() {
     OSD_ID=$(ceph osd create)
-    ceph osd crush add osd.${OSD_ID} 1 root=default
-    ceph-osd --id ${OSD_ID} --mkjournal --mkfs
-    ceph-osd --id ${OSD_ID} || ceph-osd --id ${OSD_ID} || ceph-osd --id ${OSD_ID}
+    ceph osd crush add "osd.${OSD_ID}" 1 root=default
+    ceph-osd --id "${OSD_ID}" --mkjournal --mkfs
+    ceph-osd --id "${OSD_ID}" || ceph-osd --id "${OSD_ID}" || ceph-osd --id "${OSD_ID}"
 }
 
 launch_mds() {
-    ceph auth get-or-create mds.${MDS_NAME} mon 'profile mds' mgr 'profile mds' mds 'allow *' osd 'allow *' > ${MDS_DATA}/keyring
+    ceph auth get-or-create mds.${MDS_NAME} mon 'profile mds' mgr 'profile mds' mds 'allow *' osd 'allow *' > "${MDS_DATA}/keyring"
     ceph osd pool create cephfs_data 8
     ceph osd pool create cephfs_metadata 8
     ceph fs new cephfs cephfs_metadata cephfs_data
     ceph fs ls
     ceph-mds -i ${MDS_NAME}
     ceph status
-    while [[ ! $(ceph mds stat | grep "up:active") ]]; do sleep 1; done
+    while ! ceph mds stat | grep -q "up:active"; do sleep 1; done
 }
 
 launch_mgr() {
@@ -121,7 +122,7 @@ launch_mgr() {
 
 launch_rbd_mirror() {
     ceph auth get-or-create client.rbd-mirror.${MIRROR_ID} mon 'profile rbd-mirror' osd 'profile rbd'
-    rbd-mirror --id ${MIRROR_ID} --log-file ${LOG_DIR}/rbd-mirror.log
+    rbd-mirror --id ${MIRROR_ID} --log-file "${LOG_DIR}/rbd-mirror.log"
 }
 
 launch_cephfs_mirror() {
@@ -137,8 +138,10 @@ launch_cephfs_mirror() {
 }
 
 launch_radosgw() {
-    ceph auth get-or-create client.rgw."${RGW_ID}" osd 'allow rwx' mon 'allow rw' -o ${RGW_DATA}/keyring
-    radosgw -n client.rgw."${RGW_ID}" -k ${RGW_DATA}/keyring
+    ceph auth get-or-create client.rgw."${RGW_ID}" osd 'allow rwx' mon 'allow rw' -o "${RGW_DATA}/keyring"
+    radosgw -n client.rgw."${RGW_ID}" -k "${RGW_DATA}/keyring"
+    # not going to try to make shellcheck happy with this line at this time
+    # shellcheck disable=SC2016
     timeout 60 sh -c 'until [ $(ceph -s | grep -c "rgw:") -eq 1 ]; do echo "waiting for rgw to show up" && sleep 1; done'
     radosgw-admin user create --uid admin --display-name "Admin User" --caps "buckets=*;users=*;usage=read;metadata=read" --access-key="$S3_ACCESS_KEY" --secret-key="$S3_SECRET_KEY"
 }
@@ -148,17 +151,17 @@ selftest() {
     ceph status
     test_pool=$(uuidgen)
     temp_file=$(mktemp)
-    ceph osd pool create ${test_pool} 0
-    rados --pool ${test_pool} put group /etc/group
-    rados --pool ${test_pool} get group ${temp_file}
-    diff /etc/group ${temp_file}
-    ceph osd pool delete ${test_pool} ${test_pool} --yes-i-really-really-mean-it
-    rm ${temp_file}
+    ceph osd pool create "${test_pool}" 0
+    rados --pool "${test_pool}" put group /etc/group
+    rados --pool "${test_pool}" get group "${temp_file}"
+    diff /etc/group "${temp_file}"
+    ceph osd pool delete "${test_pool}" "${test_pool}" --yes-i-really-really-mean-it
+    rm "${temp_file}"
 }
 
-FEATURESET="${CEPH_FEATURESET}"
+FEATURESET="${CEPH_FEATURESET-}"
 if [ -z "$FEATURESET" ] ; then
-    case "${CEPH_VERSION}" in
+    case "${CEPH_VERSION-}" in
         nautilus|octopus)
             FEATURESET="mon osd mgr mds rbd-mirror rgw selftest"
         ;;
@@ -169,7 +172,7 @@ if [ -z "$FEATURESET" ] ; then
 fi
 
 generate_ceph_conf
-for fname in ${FEATURESET}; do
+for fname in ${FEATURESET} ; do
     case "${fname}" in
         mon) launch_mon ;;
         osd) launch_osd ;;
@@ -186,6 +189,6 @@ for fname in ${FEATURESET}; do
     esac
 done
 
-touch ${DIR}/.ready
+touch "${DIR}/.ready"
 
 # vim: set ts=4 sw=4 sts=4 et:
